First commit of OllamaLlama3 class
add the ability to ask questions to llama3 via a local ollama setup. This is needed for the benchmark in order to validate questions for different question classes. We can also use this for many other usecases as it essentially makes local questions 'free' (meaning we don't pay per API call costs, just cost of running the server)