Adds a route for setting up the database when there is none (#64)
In order to get ourselves a fully automated end-to-end deployment going, we need a way to deploy the database schema after a new environment is stood up. This change adds that by way of a Wordpress-inspired screen. If on app-startup we detect that the database is missing, we set a special config variable that will automatically redirect a user to admin page where they can deploy the schema and migrations.

Note that at this time we don't provide capabilities to change those credentials, so they strill need to be correct when deployed.